JP-129: wip: restructuring user page
