+++ frontend/pages/user/[username].vue
@@ -2,27 +2,23 @@
 import { useTranslate } from "@tolgee/vue";
 
 const route = useRoute();
-const { t } = useTranslate();
 const router = useRouter();
-const client = useSanctumClient();
-const toast = useToast();
-const templateStore = useTemplateStore();
+const { t } = useTranslate();
 const user = useSanctumUser<User>();
-const { isAuthenticated } = useSanctumAuth();
+const { isAuthenticated } = useSanctumAuth<boolean>();
+const client = useSanctumClient();
 
-const username = ref(route.params.username);
-const displayname = ref("");
+const ALLOWED_ROUTES = ["/journey", "/dashboard?tab=templates"];
 
-const MAX_FIRST_TEMPLATES = {
-    MOBILE: 4,
-    DESKTOP: 8,
-};
+const isCurrentUser = ref<boolean>(false);
+const username = ref<string>(route.params.username as string);
+const displayname = ref<string>("");
+const profilePicture = ref("https://placehold.co/44");
+const joinDate = ref<Date | null>(null);
 
 const { data, error } = await useAsyncData("user", () =>
     client(`/api/user/${username.value}`),
 );
-
-const isCurrentUser = data.value?.username === user.value?.username;
 
 if (error.value) {
     if (error.value.statusCode === 404) {
@@ -39,409 +35,86 @@ if (error.value) {
         });
     }
 } else {
-    displayname.value = data.value.display_name;
-    username.value = data.value.username;
+    isCurrentUser.value = data.value?.username === user.value?.username;
+    displayname.value = data.value?.display_name;
+    username.value = data.value?.username;
+    //profilePicture.value = data.value?.profile_picture;
+    joinDate.value = new Date(data.value?.created_at);
     useHead({
         title: `${displayname.value} (@${username.value}) | JourneyPlanner`,
     });
 }
 
-const showMore = ref(false);
-const toggleText = ref(t.value("profile.showMore") + username.value);
-const toggleTextShort = ref(t.value("profile.showMore.short"));
-const allowedRoutes = ["/journey", "/dashboard?tab=templates"];
-const isCloseIcon = ref(false);
-const templates = ref<Template[]>([]);
-const openedTemplate = ref<Template | undefined>();
-const isTemplatePopupVisible = ref<boolean>(false);
-const moreTemplatesAvailable = ref<boolean>(true);
-const cursor = ref<string | null>(null);
-const nextCursor = ref<string | null>(null);
-const observer = ref<IntersectionObserver>();
-const loader = ref<HTMLElement | undefined>();
+const backRoute = ref("");
 
-onMounted(async () => {
+onMounted(() => {
     const lastRoute = router.options.history.state.back as string;
-    if (
-        (lastRoute &&
-            allowedRoutes.some((route) => lastRoute.startsWith(route))) ||
-        !isAuthenticated.value
-    ) {
-        isCloseIcon.value = true;
-    } else if (route?.query?.journey) {
-        isCloseIcon.value = true;
-    } else {
-        isCloseIcon.value = false;
-    }
-
-    if (
-        route.query.id &&
-        route.query.id !== "undefined" &&
-        route.query.id !== "null" &&
-        route.query.id !== ""
-    ) {
-        client(`/api/template/${route.query.id}`, {
-            async onResponse({ response }) {
-                if (response.ok) {
-                    openedTemplate.value = response._data;
-                    isTemplatePopupVisible.value = true;
-                }
-            },
-            async onResponseError({ response }) {
-                if (response.status === 404) {
-                    toast.add({
-                        severity: "error",
-                        summary: t.value("template.notfound.summary"),
-                        detail: t.value("template.notfound.summary.detail"),
-                        life: 6000,
-                    });
-                } else {
-                    toast.add({
-                        severity: "error",
-                        summary: t.value("common.toast.error.heading"),
-                        detail: t.value("common.error.unknown"),
-                        life: 6000,
-                    });
-                }
-                router.push({
-                    query: {},
-                });
-            },
-        });
-    }
-
-    watch(isTemplatePopupVisible, (value) => {
-        if (value) {
-            router.push({
-                query: {
-                    id: openedTemplate.value ? openedTemplate.value.id : null,
-                },
-            });
-        } else {
-            router.push({
-                query: {},
-            });
-        }
-    });
-});
-
-onUnmounted(() => {
-    if (observer.value && loader.value) {
-        observer.value.unobserve(loader.value);
-    }
-});
-
-const { data: templateData, refresh } = await useAsyncData(
-    "userTemplates",
-    () => client(`/api/user/${username.value}/template?cursor=${cursor.value}`),
-);
-
-watch(
-    templateData,
-    () => {
-        if (templateData.value) {
-            templates.value.push(...templateData.value.data);
-            if (templateData.value.next_cursor === null) {
-                moreTemplatesAvailable.value = false;
-            } else {
-                nextCursor.value = templateData.value.next_cursor;
-                moreTemplatesAvailable.value = true;
-            }
-        }
-    },
-    { immediate: true },
-);
-
-watch(showMore, () => {
-    if (showMore.value) {
-        if (observer.value) {
-            observer.value.disconnect();
-        }
-
-        observer.value = new IntersectionObserver((entries) => {
-            if (entries.length === 0) {
-                return;
-            }
-            const target = entries[0];
-            if (target.isIntersecting) {
-                if (moreTemplatesAvailable.value && showMore.value) {
-                    cursor.value = nextCursor.value;
-                    refresh();
-                }
-            }
-        });
-
-        if (loader.value) {
-            observer.value.observe(loader.value);
-        }
-    }
-});
-
-const firstEightTemplates = computed(() =>
-    templates.value.slice(0, MAX_FIRST_TEMPLATES.DESKTOP),
-);
-const firstFourTemplates = computed(() =>
-    templates.value.slice(0, MAX_FIRST_TEMPLATES.MOBILE),
-);
-const remainingTemplatesDesktop = computed(() =>
-    templates.value.slice(MAX_FIRST_TEMPLATES.DESKTOP),
-);
-const remainingTemplatesMobile = computed(() =>
-    templates.value.slice(MAX_FIRST_TEMPLATES.MOBILE),
-);
-
-const toggle = () => {
-    showMore.value = !showMore.value;
-    toggleText.value = showMore.value
-        ? t.value("profile.showLess") + username.value
-        : t.value("profile.showMore") + username.value;
-    toggleTextShort.value = showMore.value
-        ? t.value("profile.showLess.short")
-        : t.value("profile.showMore.short");
-};
-
-const navigateBack = () => {
-    const lastRoute = router.options.history.state.back as string;
-
-    if (!isAuthenticated.value) {
-        if (lastRoute) {
-            router.push(lastRoute);
-        } else {
-            router.push("/journey/new");
-        }
-    }
-
     if (
         lastRoute &&
-        allowedRoutes.some((route) => lastRoute.startsWith(route))
+        (ALLOWED_ROUTES.some((route) => lastRoute.startsWith(route)) ||
+            !isAuthenticated.value)
     ) {
-        router.back();
+        backRoute.value = lastRoute;
+    } else if (!isAuthenticated.value) {
+        backRoute.value = "/journey/new";
     } else if (route?.query?.journey) {
-        router.push("/journey/" + route.query.journey);
+        backRoute.value = `/journey/${route.query.journey}`;
     } else {
-        router.push("/dashboard");
+        backRoute.value = "/dashboard";
     }
-};
+});
 
-const openTemplateDialog = (template: Template) => {
-    openedTemplate.value = template;
-    isTemplatePopupVisible.value = true;
-};
-
-function removeTemplate(id: string) {
-    if (!id) return;
-    const index = templates.value.findIndex((template) => template.id === id);
-    if (index === -1) {
-        return;
+const whoseProfile = computed(() => {
+    if (isCurrentUser.value) {
+        return t.value("profile.your");
+    } else {
+        return `${displayname.value}`;
     }
-    templates.value.splice(index, 1);
-    templateStore.changeUpdate(true);
-}
+});
 </script>
 
 <template>
     <div>
-        <div class="px-1 font-nunito text-text dark:text-natural-50 md:px-3">
-            <div id="header" class="mt-5 flex flex-col">
-                <div class="flex flex-row items-center">
-                    <div class="flex flex-row items-center">
-                        <span
-                            class="mr-1 h-0.5 w-2 bg-calypso-400 xs:mr-2 xs:w-4 sm:mr-3.5 sm:w-8 md:mr-5 md:w-10"
-                        />
-                        <h1
-                            class="max-w-48 truncate text-nowrap text-xl font-medium text-text dark:text-natural-50 xs:text-2xl sm:max-w-sm md:max-w-screen-md md:text-3xl"
-                        >
-                            {{ displayname }}
-                        </h1>
-                    </div>
-                    <span
-                        class="ml-1 h-0.5 w-full bg-calypso-400 xs:ml-2 sm:ml-3.5 md:ml-5 md:mr-5"
-                    />
-                    <span
-                        v-if="isCloseIcon"
-                        class="pi pi-times cursor-pointer pl-2 pr-2 text-2xl text-natural-600 hover:text-text dark:text-natural-400 dark:hover:text-natural-50 xs:text-3xl md:pr-3 md:text-3xl lg:pr-5"
-                        @click="navigateBack"
-                    />
+        <div id="header" class="mt-5 flex flex-col">
+            <div class="flex flex-row items-center">
+                <NuxtLink
+                    :to="backRoute"
+                    class="pi pi-angle-left cursor-pointer pl-2 pr-2 text-2xl text-natural-600 hover:text-text dark:text-natural-400 dark:hover:text-natural-50 xs:text-3xl md:pr-3 md:text-3xl lg:pl-2 lg:pr-6"
+                />
+                <div class="flex w-full flex-row items-center justify-between">
+                    <h1
+                        class="max-w-48 truncate text-nowrap text-xl font-medium text-text dark:text-natural-50 xs:text-2xl sm:max-w-sm md:max-w-screen-md md:text-3xl"
+                    >
+                        {{ whoseProfile }}
+                    </h1>
                     <NuxtLink
-                        v-else
-                        to="/dashboard"
-                        class="group mx-2 flex items-center sm:ml-1 md:ml-2 lg:mr-5"
-                    >
-                        <SvgDashboardIcon
-                            class="h-7 w-7 fill-text dark:fill-natural-50 md:h-6 md:w-6"
-                        />
-                        <p
-                            class="hidden text-xl text-text group-hover:underline dark:text-natural-50 sm:block lg:text-2xl"
-                        >
-                            Dashboard
-                        </p>
-                    </NuxtLink>
-                </div>
-                <div class="mt-0.5 md:mt-2">
-                    <h2
-                        class="ml-3 max-w-48 truncate text-lg text-natural-500 dark:text-natural-300 xs:ml-6 xs:text-xl sm:ml-[2.875rem] md:ml-[3.75rem] md:max-w-screen-md md:text-2xl"
-                    >
-                        @{{ username }}
-                    </h2>
-                </div>
-            </div>
-            <div
-                class="ml-3 mt-6 pr-2 xs:ml-6 xs:pr-10 sm:ml-[2.875rem] md:ml-[3.75rem] md:mt-16 md:pr-20 lg:mt-10"
-            >
-                <h1
-                    class="text-lg font-medium text-text dark:text-natural-50 xs:text-xl md:text-2xl"
-                >
-                    <T key-name="profile.templates" />
-                    {{ username }}
-                </h1>
-                <div
-                    id="templates"
-                    class="relative mt-2 grid grid-cols-2 gap-5 sm:grid-cols-3 md:gap-4 lg:grid-cols-4 lg:gap-6"
-                >
-                    <TemplateCard
-                        v-for="template in firstEightTemplates"
-                        :key="template.id"
-                        class="hidden md:block"
-                        :template="template"
-                        :displayed-in-profile="true"
-                        :is-current-user="isCurrentUser"
-                        @open-template="openTemplateDialog(template)"
-                        @template-deleted="removeTemplate"
+                        :to="backRoute"
+                        class="pi pi-times cursor-pointer pl-2 pr-2 text-2xl text-natural-600 hover:text-text dark:text-natural-400 dark:hover:text-natural-50 xs:text-3xl md:pr-3 md:text-3xl lg:pr-10"
                     />
-                    <TemplateCardSmall
-                        v-for="template in firstFourTemplates"
-                        :key="template.id"
-                        class="md:hidden"
-                        :template="template"
-                        :displayed-in-profile="true"
-                        :is-current-user="isCurrentUser"
-                        @open-template="openTemplateDialog(template)"
-                        @template-deleted="removeTemplate"
-                    />
-                    <div
-                        v-if="firstEightTemplates.length === 0"
-                        class="col-span-full hidden text-text dark:text-natural-50 md:block"
-                    >
-                        <T key-name="template.none" />
-                    </div>
-                    <div
-                        v-if="firstFourTemplates.length === 0"
-                        class="col-span-full text-text dark:text-natural-50 md:hidden"
-                    >
-                        <T key-name="template.none" />
-                    </div>
-                </div>
-
-                <div
-                    v-if="showMore"
-                    id="more-templates"
-                    class="mt-5 grid grid-cols-2 gap-5 sm:grid-cols-3 md:gap-4 lg:grid-cols-4 lg:gap-6"
-                >
-                    <TemplateCard
-                        v-for="template in remainingTemplatesDesktop"
-                        :key="template.id"
-                        class="hidden md:block"
-                        :template="template"
-                        :is-current-user="isCurrentUser"
-                        @open-template="openTemplateDialog(template)"
-                        @template-deleted="removeTemplate"
-                    />
-                    <TemplateCardSmall
-                        v-for="template in remainingTemplatesMobile"
-                        :key="template.id"
-                        class="md:hidden"
-                        :template="template"
-                        :displayed-in-profile="true"
-                        :is-current-user="isCurrentUser"
-                        @open-template="openTemplateDialog(template)"
-                        @template-deleted="removeTemplate"
-                    />
-                </div>
-                <div ref="loader" class="col-span-full">
-                    <div v-if="moreTemplatesAvailable && showMore">
-                        <div class="flex justify-center">
-                            <ProgressSpinner class="w-10" />
-                        </div>
-                        <div
-                            class="flex justify-center italic text-text dark:text-natural-50"
-                        >
-                            <T key-name="dashboard.templates.loading" />
-                        </div>
-                    </div>
-                </div>
-                <div
-                    v-if="remainingTemplatesDesktop.length > 0"
-                    class="mt-4 flex justify-center max-md:hidden"
-                >
-                    <button
-                        class="flex flex-col items-center justify-center text-text dark:text-natural-50"
-                        @click="toggle"
-                    >
-                        <span>{{ toggleText }}</span>
-                        <span
-                            class="pi mt-1"
-                            :class="
-                                showMore
-                                    ? 'pi-chevron-up order-first mb-1'
-                                    : 'pi-chevron-down'
-                            "
-                        />
-                    </button>
-                </div>
-                <div
-                    v-if="remainingTemplatesMobile.length > 0"
-                    class="mt-4 flex justify-center md:hidden"
-                >
-                    <button
-                        class="flex flex-col items-center justify-center text-text dark:text-natural-50"
-                        @click="toggle"
-                    >
-                        <span>{{ toggleTextShort }}</span>
-                        <span
-                            class="pi mt-1"
-                            :class="
-                                showMore
-                                    ? 'pi-chevron-up order-first mb-1'
-                                    : 'pi-chevron-down'
-                            "
-                        />
-                    </button>
                 </div>
             </div>
         </div>
-        <div id="dialogs">
-            <TemplatePopup
-                v-if="openedTemplate"
-                :template="openedTemplate"
-                :is-template-dialog-visible="isTemplatePopupVisible"
-                @close="
-                    isTemplatePopupVisible = false;
-                    openedTemplate = undefined;
-                "
-            />
-            <ConfirmDialog
-                :draggable="false"
-                close-on-escape
-                dismissable-mask
-                group="username"
-                :pt="{
-                    header: {
-                        class: 'bg-natural-50 dark:bg-natural-900 text-text dark:text-natural-50 font-nunito',
-                    },
-                    content: {
-                        class: 'bg-natural-50 dark:bg-natural-900 text-text dark:text-natural-50 font-nunito',
-                    },
-                    footer: {
-                        class: 'bg-natural-50 dark:bg-natural-900 text-text dark:text-natural-50 font-nunito gap-x-5',
-                    },
-                    closeButton: {
-                        class: 'bg-natural-50 dark:bg-natural-900 text-natural-500 hover:text-text dark:text-natural-400 hover:dark:text-natural-50 font-nunito',
-                    },
-                    closeButtonIcon: {
-                        class: 'h-5 w-5',
-                    },
-                }"
-            />
+        <div id="content" class="mt-5 flex w-full flex-row pl-16">
+            <div
+                id="profile"
+                class="flex h-[65vh] w-[48vh] flex-col items-center rounded-lg border-[3px] border-calypso-400 pt-5"
+            >
+                <NuxtImg
+                    :src="profilePicture"
+                    class="h-40 w-40 rounded-full object-contain"
+                    :alt="t('profile.picture')"
+                />
+                <h2 class="mt-6 text-2xl">{{ displayname }}</h2>
+                <h3 class="mt-1 text-xl text-natural-800">@{{ username }}</h3>
+                <span class="mb-1 mt-auto"
+                    ><T
+                        key-name="profile.created_at"
+                        :params="{
+                            year: joinDate ? joinDate.getFullYear() : '',
+                        }"
+                /></span>
+            </div>
+            <div id="templates"></div>
         </div>
     </div>
 </template>
